Registration update

+++ app/Http/Livewire/Dashboard/CategoryLivewire.php
@@ -5,6 +5,7 @@ namespace App\Http\Livewire\Dashboard;
 use Livewire\Component;
 use App\Models\Mainmenu;
 use App\Models\Categories;
+use Illuminate\Support\Str;
 use Livewire\WithPagination;
 use Livewire\WithFileUploads;
 use Illuminate\Support\Facades\Auth;
@@ -132,7 +133,11 @@ class CategoryLivewire extends Component
                     $croppedImage = base64_decode(preg_replace('#^data:image/\w+;base64,#i', '', $this->tempImg));
 
                     if( $this->imgReader){
-                        Storage::disk('s3')->delete($this->imgReader);
+                        if (Str::startsWith($this->imgReader, 'mine-setting/')) {
+                            //Do nothing
+                        } else {
+                            Storage::disk('s3')->delete($this->imgReader);
+                        }
                         Storage::disk('s3')->put($this->objectName, $croppedImage);
                     } else {
                         Storage::disk('s3')->put($this->objectName, $croppedImage);
@@ -149,7 +154,11 @@ class CategoryLivewire extends Component
                 $croppedImageCover = base64_decode(preg_replace('#^data:image/\w+;base64,#i', '',  $this->tempImgCover));
                 try {
                     if( $this->imgReaderCover){
-                        Storage::disk('s3')->delete($this->imgReaderCover);
+                        if (Str::startsWith($this->imgReaderCover, 'mine-setting/')) {
+                            //Do nothing
+                        } else {
+                            Storage::disk('s3')->delete($this->imgReaderCover);
+                        }
                         Storage::disk('s3')->put($this->objectNameCover, $croppedImageCover);
                     } else {
                         Storage::disk('s3')->put($this->objectNameCover, $croppedImageCover);
@@ -237,7 +246,7 @@ class CategoryLivewire extends Component
             $this->status = $menu_edit->status;
             $this->priority = $menu_edit->priority;
             $this->imgReader = $menu_edit->img;
-            $this->imgReaderCover = $menu_edit->cover;
+            $this->imgReaderCover = $menu_edit->cover ?? null;
 
             $this->old_category_data = [
                 'id' => $menu_edit->id,
@@ -274,7 +283,12 @@ class CategoryLivewire extends Component
                     $croppedImage = base64_decode(preg_replace('#^data:image/\w+;base64,#i', '', $this->tempImg));
 
                     if( $this->imgReader){
-                        Storage::disk('s3')->delete($this->imgReader);
+                        // dd($this->imgReader,app('cloudfront').'mine-setting/');
+                        if (Str::startsWith($this->imgReader, 'mine-setting/')) {
+                            //Do nothing
+                        } else {
+                            Storage::disk('s3')->delete($this->imgReader);
+                        }
                         Storage::disk('s3')->put($this->objectName, $croppedImage);
                     } else {
                         Storage::disk('s3')->put($this->objectName, $croppedImage);
@@ -291,7 +305,11 @@ class CategoryLivewire extends Component
                 $croppedImageCover = base64_decode(preg_replace('#^data:image/\w+;base64,#i', '',  $this->tempImgCover));
                 try {
                     if( $this->imgReaderCover){
-                        Storage::disk('s3')->delete($this->imgReaderCover);
+                        if (Str::startsWith($this->imgReaderCover, 'mine-setting/')) {
+                            //Do nothing
+                        } else {
+                            Storage::disk('s3')->delete($this->imgReaderCover);
+                        }
                         Storage::disk('s3')->put($this->objectNameCover, $croppedImageCover);
                     } else {
                         Storage::disk('s3')->put($this->objectNameCover, $croppedImageCover);
@@ -453,9 +471,18 @@ class CategoryLivewire extends Component
         try{
             if ($this->confirmDelete && $this->categoryNameToDelete === $this->showTextTemp) {
                 Categories::find($this->category_selected_id_delete->id)->delete();
-                Storage::disk('s3')->delete($this->category_selected_id_delete->img);
+                if (Str::startsWith($this->category_selected_id_delete->img, 'mine-setting/')) {
+                    //Do nothing
+                } else {
+                    Storage::disk('s3')->delete($this->category_selected_id_delete->img);
+                }
+
                 if($this->category_selected_id_delete->cover){
-                    Storage::disk('s3')->delete($this->category_selected_id_delete->cover);
+                    if (Str::startsWith($this->category_selected_id_delete->cover, 'mine-setting/')) {
+                        //Do nothing
+                    } else {
+                        Storage::disk('s3')->delete($this->category_selected_id_delete->cover);
+                    }
                 }
                 $this->category_selected_id_delete = null;
                 $this->category_selected_name_delete = null;
